requirements.txt
cupy_cuda12x==13.3.0
opencv_contrib_python==4.8.0.76
opencv_python==4.8.0.76
scipy
ultralytics
h5py
lap
numpy==1.26.2
pyqt5
pyinstaller 
pytorch==2.4.1
onnxruntime===1.19.2

#  system library
# multiprocessing, sys, queue, time, os, json, ...